MIT Class 2016 Big Data & Social Analytics Certificate
Module 1: Foundations of big data and social physics
Work through the units in this module to engage with and apply knowledge
related to big data, social physics, data quality, and data privacy, and
complete your practical assessment.

Module 2: Personal sensors and human behavior
This module will explore the various sources of data, how to leverage
personal sensors to predict behavior, and introduce you to the key
features of Funf.

Module 3: First-order analysis and data exploration
This module will explore the basics of data dynamics and data
interrogation, and the importance of interdisciplinary research and
testing on yourself, to assist in analysing a data set.

Module 4: Networks of physical interactions
This module will explore the fundamentals and structure of peer
networks, and introduce you to the concepts of network theory and graph
clustering.

Module 5: Second-order analysis and data exploration
This module will introduce data exploration and visualization techniques
by using the Bandicoot application.

Module 6: Using data to affect behavior change
This module will explore how to affect behavior change through social
influence and social incentives, as well as the importance of
privacy-protecting practices.

Module 7: Application of big data in industry
This module will explore how to use data analysis to effect a desired
change, and apply big data in various industries.

Module 8: Data in action
This module will explore how to understand, apply, and gain insights
into the use of data in action.